fix(newsletter): confirm via Server Component page — evita GET mutation
